Fix SAS token escaping in CopyRequest
